Rename 'TokenSchema' -> 'JwtSchema' and move it to jwt module package
